stabilize test fixtures and room member writes

+++ crates/lilium-models/src/ingestion/mod.rs
@@ -23,7 +23,7 @@ pub struct EventEnvelope {
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct EventProcessorOffset {
     pub processor_id: String,
-    pub last_processed_id: i64,
+    pub last_processed_id: i32,
     pub last_processed_timestamp: Option<DateTime<Utc>>,
     pub last_processed_at: Option<DateTime<Utc>>,
     pub updated_at: DateTime<Utc>,
